Replaced the usage of 'RealityFoundation'
The Xcode 13 beta exposed the framework 'RealityFoundation'.
The initial version of this library made use of this split
of 'RealityKit' and 'RealityFoundation' features.

This commit removes the usage of 'RealityFoundation' to provide
a seamless backward compatibility with older Xcode/Swift versions.

Follow-up: Remove the split from the '+Simulator.swift' files.