Tests: Reworked the tests to leverage Xcode 12.5 and macOS 11
Previously, most (unit) tests weren't executing the code but
solely created a closure that contained the test code.

This was done because the CI/CD was running macOS 10.x and
an older version of Xcode where the tests caused crashes
because they created (Metal) objects which weren't supported
by the iOS Simulator(s) before macOS 11.